feat: 접근 제어자 수정 (#8)
